# openapi.yaml
openapi: 3.0.3
info:
  title: Serverless Sudoku API
  version: 1.0.0
  description: >
    The API contract for the Serverless Sudoku application. 
    This API is completely stateless. The client is responsible for maintaining the current board state.
servers:
  - url: http://localhost:8080/api/v1
    description: Local Quarkus Development Server
  - url: https://api.your-amplify-domain.com/v1
    description: Production AWS API Gateway
tags:
  - name: Health
    description: Health check endpoint
  - name: Puzzles
    description: Endpoints for generating and validating Sudoku grids
  - name: Games
    description: Endpoints for managing game state
  - name: Players
    description: Endpoints for player profile management

security:
  - CognitoJWT: []

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns the current health status of the service.
      security: []
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'


  /puzzles/generate:
    get:
      tags:
        - Puzzles
      summary: Generate a new Sudoku puzzle
      description: Returns a newly generated 9x9 Sudoku grid based on the requested difficulty.
      security: []
      parameters:
        - in: query
          name: difficulty
          schema:
            $ref: '#/components/schemas/Difficulty'
          required: false
          description: The difficulty level of the puzzle. Defaults to 'medium'.
      responses:
        '200':
          description: A new Sudoku puzzle generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PuzzleResponse'

  /puzzles/validate:
    post:
      tags:
        - Puzzles
      security: []
      summary: Validate a Sudoku board
      description: Checks the provided 9x9 grid against Sudoku rules. Returns true if valid, or a list of conflicting coordinates if invalid.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardRequest'
      responses:
        '200':
          description: Validation completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '400':
          description: Malformed board payload (e.g., not 9x9).

  /puzzles/hint:
    post:
      tags:
        - Puzzles
      security: []
      summary: Request a logical learning hint for the current board
      description: Evaluates the board using human logic techniques. Returns a progressive hint (Nudge, Focus, Reveal) and the exact board modifications required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardRequest'
      responses:
        '200':
          description: A logical hint was found successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HintResponse'
        '404':
          description: No logical next step found (board might be invalid or requires guessing, which we want to avoid).

  /puzzles/candidates:
    post:
      tags:
        - Puzzles
      security: []
      summary: Auto-calculate candidates for the board
      description: Evaluates the current board and returns a grid of all valid candidate numbers for every empty cell.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardRequest'
      responses:
        '200':
          description: Candidates calculated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidatesResponse'

  /games:
    post:
      tags:
        - Games
      summary: Start a new game
      description: Generates a new puzzle, creates a database record, and returns the gameId.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                difficulty:
                  $ref: '#/components/schemas/Difficulty'
      responses:
        '201':
          description: Game created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'

  /games/from-image:
    post:
      tags:
        - Games
      summary: Start a new game from a scanned image grid
      description: Creates a new game using a grid extracted from an image (e.g., photographed puzzle). Returns the full game state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameFromGridRequest'
      responses:
        '201':
          description: Game created successfully from image grid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'

  /games/current:
    get:
      tags:
        - Games
      summary: Get the current in-progress game
      description: Retrieves the authenticated user's current in-progress game, if one exists.
      responses:
        '200':
          description: Current game state retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'
        '204':
          description: No current game in progress.

  /games/{gameId}:
    get:
      tags:
        - Games
      summary: Load a saved game
      description: Retrieves the full game state so the user can resume.
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Game state retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'
        '404':
          description: Game not found.

    patch:
      tags:
        - Games
      summary: Auto-save game progress
      description: Updates the current grid, notes, and time spent.
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameUpdateRequest'
      responses:
        '200':
          description: Game saved successfully.


  /players/me:
    get:
      tags:
        - Players
      summary: Get the current player's profile
      description: Returns the authenticated user's profile, including display name and account timestamps.
      responses:
        '200':
          description: Player profile retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerProfile'
        '404':
          description: Player profile not found.

components:
  securitySchemes:
    CognitoJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT access token.

  schemas:
    GameState:
      type: object
      required:
        - userId
        - gameId
        - difficulty
        - originalGrid
        - currentGrid
        - candidates
        - timeSpentSeconds
        - status
      properties:
        userId:
          type: string
          description: The ID of the authenticated user who owns this game.
        gameId:
          type: string
          format: uuid
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        originalGrid:
          $ref: '#/components/schemas/Grid'
        currentGrid:
          $ref: '#/components/schemas/Grid'
        candidates:
          $ref: '#/components/schemas/CandidatesGrid'
        timeSpentSeconds:
          type: integer
          description: Total time the user has actively spent on this puzzle.
        status:
          type: string
          enum: [IN_PROGRESS, SOLVED, ABANDONED]
      example:
        gameId: "123e4567-e89b-12d3-a456-426614174000"
        difficulty: "medium"
        originalGrid: [[5,3,0...]] # Truncated for brevity
        currentGrid: [[5,3,4...]] 
        candidates: [[[], [], [1,2]]] 
        timeSpentSeconds: 145
        status: "IN_PROGRESS"

    GameUpdateRequest:
      type: object
      required:
        - currentGrid
        - candidates
        - timeSpentSeconds
      properties:
        currentGrid:
          $ref: '#/components/schemas/Grid'
        candidates:
          $ref: '#/components/schemas/CandidatesGrid'
        timeSpentSeconds:
          type: integer
        isComplete:
          type: boolean
          description: Pass true if the user just filled the final correct square.

    Difficulty:
      type: string
      enum: [easy, medium, hard, expert]
      default: medium
      example: "medium"

    Grid:
      type: array
      description: A 9x9 matrix representing the Sudoku board. 0 represents an empty cell.
      minItems: 9
      maxItems: 9
      items:
        type: array
        minItems: 9
        maxItems: 9
        items:
          type: integer
          minimum: 0
          maximum: 9

    PuzzleResponse:
      type: object
      required:
        - originalGrid
        - difficulty
      properties:
        originalGrid:
          $ref: '#/components/schemas/Grid'
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        minRank:
          type: integer
          description: The minimum strategy rank used when generating this puzzle.

    BoardRequest:
      type: object
      required:
        - currentGrid
      properties:
        currentGrid:
          $ref: '#/components/schemas/Grid'
        minRank:
          type: integer
          description: Optional minimum strategy rank to filter hints (hints below this rank are skipped).
        excludedRanks:
          type: array
          description: Optional list of strategy ranks to exclude from hint selection.
          items:
            type: integer
      example:
        currentGrid: [
          [5,3,0,0,7,0,0,0,0],
          [6,0,0,1,9,5,0,0,0],
          [0,9,8,0,0,0,0,6,0],
          [8,0,0,0,6,0,0,0,3],
          [4,0,0,8,0,3,0,0,1],
          [7,0,0,0,2,0,0,0,6],
          [0,6,0,0,0,0,2,8,0],
          [0,0,0,4,1,9,0,0,5],
          [0,0,0,0,8,0,0,7,9]
        ]

    ValidationResponse:
      type: object
      required:
        - isValid
      properties:
        isValid:
          type: boolean
          description: True if the board violates no Sudoku rules.
        isSolved:
          type: boolean
          description: True if the board is completely filled and valid.
        errors:
          type: array
          description: List of coordinates that are conflicting (if isValid is false).
          items:
            $ref: '#/components/schemas/Coordinate'

    HintResponse:
      type: object
      required:
        - techniqueName
        - markdownSlug
        - difficulty
        - strategyRank
        - nudge
        - focus
        - reveal
        - highlightCells
        - eliminatedCandidates
        - solvedCells
        - focusCandidates
      properties:
        techniqueName:
          type: string
          description: The friendly name of the logical technique used.
        markdownSlug:
          type: string
          description: The identifier linking to the frontend markdown file (e.g., "naked-pair").
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        strategyRank:
          type: integer
          description: Numeric rank of the strategy used (lower = simpler technique).
        nudge:
          type: string
          description: Level 1 Hint - A gentle push toward the technique.
        focus:
          type: string
          description: Level 2 Hint - Points the user to the specific area of the board.
        reveal:
          type: string
          description: Level 3 Hint - Explains exactly what to do and why.
        highlightCells:
          type: array
          description: Coordinates of cells that form the pattern (e.g., the two cells of a Naked Pair).
          items:
            $ref: '#/components/schemas/Coordinate'
        eliminatedCandidates:
          type: array
          description: Pencil marks that should be removed based on this logic.
          items:
            $ref: '#/components/schemas/CoordinateCandidate'
        solvedCells:
          type: array
          description: Cells that can be definitively filled based on this logic.
          items:
            $ref: '#/components/schemas/ActionableCell'
        focusCandidates:
          type: array
          description: Candidate marks to visually highlight as part of the hint focus area (without eliminating them).
          items:
            $ref: '#/components/schemas/CoordinateCandidate'
      example:
        techniqueName: "Naked Pair"
        markdownSlug: "naked-pair"
        difficulty: "medium"
        nudge: "There is a Naked Pair hiding somewhere on the board."
        focus: "Look closely at the candidates in Row 4."
        reveal: "Cells (4,2) and (4,6) only contain candidates 3 and 7. They form a Naked Pair, meaning 3 and 7 can be safely removed from all other cells in Row 4."
        highlightCells:
          - row: 4
            col: 2
          - row: 4
            col: 6
        eliminatedCandidates:
          - row: 4
            col: 8
            value: 3
          - row: 4
            col: 9
            value: 7
        solvedCells: []

    Coordinate:
      type: object
      required:
        - row
        - col
      properties:
        row:
          type: integer
          minimum: 0
          maximum: 8
        col:
          type: integer
          minimum: 0
          maximum: 8
      example:
        row: 4
        col: 2
    
    CandidatesResponse:
      type: object
      required:
        - candidatesGrid
      properties:
        candidatesGrid:
          $ref: '#/components/schemas/CandidatesGrid'

    CandidatesGrid:
      type: array
      description: A 9x9 matrix where each cell contains an array of possible valid integers (1-9). If a cell is already filled in the main board, its candidate array is empty.
      minItems: 9
      maxItems: 9
      items:
        type: array
        minItems: 9
        maxItems: 9
        items:
          type: array
          description: List of valid candidate numbers for this specific cell.
          items:
            type: integer
            minimum: 1
            maximum: 9

    CoordinateCandidate:
      type: object
      required:
        - row
        - col
        - value
      properties:
        row:
          type: integer
          minimum: 0
          maximum: 8
        col:
          type: integer
          minimum: 0
          maximum: 8
        value:
          type: integer
          minimum: 1
          maximum: 9
          description: The candidate number to remove.
      example:
        row: 4
        col: 8
        value: 3 

    ActionableCell:
      type: object
      required:
        - row
        - col
        - value
      properties:
        row:
          type: integer
          minimum: 0
          maximum: 8
        col:
          type: integer
          minimum: 0
          maximum: 8
        value:
          type: integer
          minimum: 1
          maximum: 9
          description: The correct number to place in this cell.
      example:
        row: 2
        col: 4
        value: 7

    HealthResponse:
      type: object
      required:
        - status
        - service
      properties:
        status:
          type: string
          description: Health status of the service (e.g., "UP").
          example: "UP"
        service:
          type: string
          description: Name of the service.
          example: "sudoku-api"

    PlayerProfile:
      type: object
      required:
        - userId
        - email
        - displayName
        - createdAt
        - updatedAt
      properties:
        userId:
          type: string
          description: Unique identifier for the player (from Cognito sub claim).
        email:
          type: string
          format: email
          description: The player's email address.
        displayName:
          type: string
          description: The player's display name.
        createdAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the player profile was created.
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the player profile was last updated.

    CreateGameFromGridRequest:
      type: object
      required:
        - originalGrid
      properties:
        originalGrid:
          $ref: '#/components/schemas/Grid'
      description: Request to create a game from a pre-existing grid (e.g., scanned from an image).